refactor: consolidate Twilio signature verification into shared module
