+++ gateway/package.json
@@ -2,6 +2,9 @@
   "name": "@vellumai/vellum-gateway",
   "version": "0.4.6",
   "type": "module",
+  "exports": {
+    "./twilio/verify": "./src/twilio/verify.ts"
+  },
   "scripts": {
     "dev": "bun run --watch src/index.ts",
     "dev:proxy": "GATEWAY_RUNTIME_PROXY_ENABLED=true GATEWAY_RUNTIME_PROXY_REQUIRE_AUTH=false bun run --watch src/index.ts",
